Add interactive mode
This allows modification of lines and general verification that it's
doing what you expect.  It also importantly adds the ability to make the
sort of tweaks that are often necessary: making up new entries for the
frames and partial frames at the start of a roll that you didn't record
an entry for because you thought you were still winding through
light-burnt film.